# [PAL/Linux-SGX] Add exitless system calls (new version)