--- a/Pal/src/host/Linux-SGX/enclave_framework.c
+++ b/Pal/src/host/Linux-SGX/enclave_framework.c
@@ -48,14 +48,19 @@ void* sgx_prepare_ustack(void) {
     return old_ustack;
 }
 
-void* sgx_alloc_on_ustack(uint64_t size) {
+void* sgx_alloc_on_ustack_aligned(uint64_t size, size_t alignment) {
+    assert(IS_POWER_OF_2(alignment));
     void* ustack = GET_ENCLAVE_TLS(ustack) - size;
-    ustack = ALIGN_DOWN_PTR(ustack, 16);
+    ustack = ALIGN_DOWN_PTR_POW2(ustack, alignment);
     if (!sgx_is_completely_outside_enclave(ustack, size)) {
         return NULL;
     }
     SET_ENCLAVE_TLS(ustack, ustack);
     return ustack;
+}
+
+void* sgx_alloc_on_ustack(uint64_t size) {
+    return sgx_alloc_on_ustack_aligned(size, 1);
 }
 
 void* sgx_copy_to_ustack(const void* ptr, uint64_t size) {