--- a/Pal/src/host/Linux-SGX/sgx_api.h
+++ b/Pal/src/host/Linux-SGX/sgx_api.h
@@ -26,6 +26,7 @@ bool sgx_is_completely_within_enclave(const void* addr, uint64_t size);
 bool sgx_is_completely_outside_enclave(const void* addr, uint64_t size);
 
 void* sgx_prepare_ustack(void);
+void* sgx_alloc_on_ustack_aligned(uint64_t size, size_t alignment);
 void* sgx_alloc_on_ustack(uint64_t size);
 void* sgx_copy_to_ustack(const void* ptr, uint64_t size);
 void sgx_reset_ustack(const void* old_ustack);