… Documentation/manifest-syntax.rst
… LibOS/shim/test/regression/multi_pthread_exitless.manifest.template
… LibOS/shim/test/regression/test_libos.py
… Pal/src/host/Linux-SGX/enclave_ocalls.c
… Pal/src/host/Linux-SGX/sgx_main.c