@dimakuv 2019-11-22
## Performance numbers

Quick performance evaluation. I used an SGX v1 machine with Intel(R) Xeon(R) CPU E3-1280 v5 @ 3.70GHz, 8 logical CPUs, 8MB LLC cache, and 48GB of RAM, with Ubuntu 18.04 and Linux 5.1.

**TLDR**: Exitless improves performance **~2X**.

### Redis

Redis is (almost) single-threaded. Our example manifest creates 8 enclave threads, so I added 8 RPC threads. Most of them sleep anyway and do not affect performance.

Running our example like this: `$ SGX=1 ./pal_loader redis-server --save '' --protected-mode no`.

- without Exitless:
```
$ src/src/redis-benchmark -t get,set -q
SET: 11018.07 requests per second
GET: 11181.93 requests per second
```

- with Exitless:
```
$ src/src/redis-benchmark -t get,set -q
SET: 24636.61 requests per second
GET: 25000.00 requests per second
```

### Memcached

Memcached is weirdly multi-threaded. Even when running with 2 threads, it creates another ~8 threads for some bookkeeping. Thus, I added 10 RPC threads. The resulting configuration occupies all my 8 logical CPUs, each with almost 100% utilization.

Running our example like this: `$ SGX=1 ./pal_loader memcached -u nobody -t 2`. I restricted to two threads since I only have a 8-logical CPUs machine.

- without Exitless:
```
$ memtier_benchmark --port=11211 --protocol=memcache_binary --hide-histogram
Type         Ops/sec     Hits/sec   Misses/sec      Latency       KB/sec
-------------------------------------------------------------------------
Sets         2570.12          ---          ---     45.21000       197.94
Gets        25673.00         0.00     25673.00      0.17000       975.00
Totals      28243.13         0.00     25673.00      4.26900      1172.95
```

- with Exitless:
```
$ memtier_benchmark --port=11211 --protocol=memcache_binary --hide-histogram
Type         Ops/sec     Hits/sec   Misses/sec      Latency       KB/sec
-------------------------------------------------------------------------
Sets         4946.72          ---          ---     16.27800       380.98
Gets        49412.89         0.00     49412.89      1.52600      1876.59
Totals      54359.61         0.00     49412.89      2.86800      2257.57
```