@dimakuv 2020-03-20
Jenkins, retest Jenkins-18.04 please (`futex_wait_bitset01` LTP test failed)